+++ conanfile.py
@@ -32,6 +32,7 @@ class BeezConan(ConanFile):
         self.requires("rapidxml/1.13")
         self.requires("fast-cpp-csv-parser/cci.20240102")
         self.requires("tomlplusplus/3.4.0")
+        self.requires("libcurl/8.12.1")
 
     def layout(self):
         cmake_layout(self)